creating terminal graphical example_client_2
lots of TODOs.
beginning to allow some punctuation.
you can use "D" for "done".
you can use "B" to delete the previous word, useful for when you
misspell a word and dont want to leave the new state to go back
and fix it.

tests may not work, and new tests need to be written
but for my day-to-day use, it seems to work fine for now.